third_party/WebKit/LayoutTests/virtual/color_space/fast/canvas/color-space/crasher.html

<!DOCTYPE HTML>
<script src="../../../../../resources/testharness.js"></script>
<script src="../../../../../resources/testharnessreport.js"></script>
<script>

// The reference values are generated by calling SkImage::makeColorSpace()
// in a Skia fiddle as this is the API used in ImageBitmap for color conversion.
// SkColorSpaceXform() may generate slightly different results.
// Please see: https://fiddle.skia.org/c/94578944d9d52c2b4c2cadda88a4e204

var p3Red =   [0.823242, 0.031372, 0.015686, 1]; // SRGB(255,0,0,255)
var p3Green = [0.176392, 0.968262, 0.070557, 1]; // SRGB(0,255,0,255)
var p3Blue =  [0, 0, 0.909668, 1];               // SRGB(0,0,255,255)
var p3Black = [0, 0, 0, 1];                      // SRGB(0,0,0,255)

var p3OpaqueRed = [0.270508, 0.019608, 0.015686, 1];   // SRGB(155,27,27,255)
var p3OpaqueGreen = [0.066650, 0.317627, 0.035278, 1]; // SRGB(27,155,27,255)
var p3OpaqueBlue = [0.011765, 0.011765, 0.297852, 1];  // SRGB(27,27,155,255)
var p3OpaqueBlack = [0.011765, 0.011765, 0.011765, 1]; // SRGB(27,27,27,255)

// SRGB(155,27,27,128)
var p3TransparentRed = [0.270508, 0.019608, 0.015686, 0.501953];
// SRGB(27,155,27,128)
var p3TransparentGreen = [0.066650, 0.317627, 0.035278, 0.501953];
// SRGB(27,27,155,128)
var p3TransparentBlue = [0.011765, 0.011765, 0.297852, 0.501953];
// SRGB(27,27,27,128)
var p3TransparentBlack = [0.011765, 0.011765, 0.011765, 0.501953];

function testPixels(ctx, tests, sourceType)
{
    var actual, expected, tolerance = 0.03;
    if (sourceType == 'video')
        tolerance = 0.03;
    for (var i = 0; i < tests.length; i++) {
        actual = ctx.getImageData(tests[i][0], tests[i][1], 1, 1).dataUnion;
        expected = tests[i][2];
        assert_true(actual.length === expected.length);
       for (var i = 0; i < actual.length; i++)
           assert_approx_equals(actual[i], expected[i], tolerance, tests[i][3]);
    }
}

function checkNoCrop(imageBitmap, colorInfo, sourceType)
{
    var canvas = document.createElement('canvas');
    canvas.width = 5;
    canvas.height = 5;
    var ctx = canvas.getContext('2d',
        {colorSpace: 'p3', pixelFormat:'float16'});
    ctx.clearRect(0, 0, canvas.width, canvas.height);
    ctx.drawImage(imageBitmap, 0, 0);
    var tests;
    if (colorInfo == 'fullColor')
        tests = [[0, 0, p3Red, "This pixel should be P3 red."],
                 [3, 0, p3Green, "This pixel should be P3 green."],
                 [0, 3, p3Blue, "This pixel should be P3 blue."],
                 [3, 3, p3Black, "This pixel should be P3 black."],
                 [4, 4, p3Black, "This pixel should be P3 black."]];
    else if (colorInfo == 'opaque')
        tests = [[0, 0, p3OpaqueRed, "This pixel should be P3 like red."],
                 [3, 0, p3OpaqueGreen, "This pixel should be P3 like green."],
                 [0, 3, p3OpaqueBlue, "This pixel should be P3 like blue."],
                 [3, 3, p3OpaqueBlack, "This pixel should be P3 like black."],
                 [4, 4, p3Black, "This pixel should be P3 black."]];
    else if (colorInfo == 'transparent')
        tests = [[0, 0, p3TransparentRed,
                     "This pixel should be P3 transparent red."],
                 [3, 0, p3TransparentGreen,
                     "This pixel should be P3 transparent green."],
                 [0, 3, p3TransparentBlue,
                     "This pixel should be P3 transparent blue."],
                 [3, 3, p3TransparentBlack,
                     "This pixel should be P3 transparent black."],
                 [4, 4, p3Black, "This pixel should be P3 black."]];
    testPixels(ctx, tests, sourceType);
}

function checkCrop(imageBitmap, colorInfo, sourceType)
{
    var canvas = document.createElement('canvas');
    canvas.width = 5;
    canvas.height = 5;
    var ctx = canvas.getContext('2d',
        {colorSpace: 'p3', pixelFormat:'float16'});
    ctx.clearRect(0, 0, canvas.width, canvas.height);
    ctx.drawImage(imageBitmap, 0, 0);
    var tests;
    if (colorInfo === 'fullColor')
        tests = [[0, 0, p3Red, "This pixel should be P3 red."],
                 [1, 0, p3Green, "This pixel should be P3 green."],
                 [0, 1, p3Blue, "This pixel should be P3 blue."],
                 [1, 1, p3Black, "This pixel should be P3 black."],
                 [2, 2, p3Black, "This pixel should be P3 black."]];
    else if (colorInfo === 'opaque')
        tests = [[0, 0, p3OpaqueRed, "This pixel should be P3 like red."],
                 [1, 0, p3OpaqueGreen, "This pixel should be P3 like green."],
                 [0, 1, p3OpaqueBlue, "This pixel should be P3 like blue."],
                 [1, 1, p3OpaqueBlack, "This pixel should be P3 like black."],
                 [2, 2, p3Black, "This pixel should be P3 black."]];
    else if (colorInfo === 'transparent')
        tests = [[0, 0, p3TransparentRed,
                     "This pixel should be P3 transparent red."],
                 [1, 0, p3TransparentGreen,
                     "This pixel should be P3 transparent green."],
                 [0, 1, p3TransparentBlue,
                     "This pixel should be P3 transparent blue."],
                 [1, 1, p3TransparentBlack,
                     "This pixel should be P3 transparent black."],
                 [2, 2, p3Black, "This pixel should be P3 black."]];
    testPixels(ctx, tests, sourceType);
}

function compareBitmaps(bitmap1, bitmap2)
{
    var canvas1 = document.createElement('canvas');
    var canvas2 = document.createElement('canvas');
    canvas1.width = 5;
    canvas1.height = 5;
    canvas2.width = 5;
    canvas2.height = 5;
    var ctx1 = canvas1.getContext('2d',
        {colorSpace: 'p3', pixelFormat:'float16'});
    var ctx2 = canvas2.getContext('2d',
        {colorSpace: 'p3', pixelFormat:'float16'});
    ctx1.clearRect(0, 0, canvas1.width, canvas1.height);
    ctx2.clearRect(0, 0, canvas2.width, canvas2.height);
    ctx1.drawImage(bitmap1, 0, 0);
    ctx2.drawImage(bitmap2, 0, 0);
    var data1 = ctx1.getImageData(0, 0, 5, 5).dataUnion;
    var data2 = ctx2.getImageData(0, 0, 5, 5).dataUnion;
    var dataMatched = true;
    for (var i = 0; i < data1.length; i++) {
        if (data1[i] != data2[i]) {
            dataMatched = false;
            break;
        }
    }
    assert_false(dataMatched);
}

function testImageBitmap(source, colorInfo, sourceType)
{
    return Promise.all([
        createImageBitmap(source, {colorSpaceConversion: "p3", resizeWidth: 4,
            resizeHeight: 4, resizeQuality: "high"}),
    //     createImageBitmap(source, {colorSpaceConversion: "p3", resizeWidth: 4,
    //         resizeHeight: 4, resizeQuality: "medium"}),
    //     createImageBitmap(source, {colorSpaceConversion: "p3", resizeWidth: 4,
    //         resizeHeight: 4, resizeQuality: "low"}),
    //     createImageBitmap(source, {colorSpaceConversion: "p3", resizeWidth: 4,
    //         resizeHeight: 4, resizeQuality: "pixelated"}),
    //     createImageBitmap(source, 5, 5, 10, 10, {colorSpaceConversion: "p3",
    //         resizeWidth: 20, resizeHeight: 20, resizeQuality: "high"}),
    //     createImageBitmap(source, 5, 5, 10, 10, {colorSpaceConversion: "p3",
    //         resizeWidth: 20, resizeHeight: 20, resizeQuality: "medium"}),
    //     createImageBitmap(source, 5, 5, 10, 10, {colorSpaceConversion: "p3",
    //         resizeWidth: 20, resizeHeight: 20, resizeQuality: "low"}),
    //     createImageBitmap(source, 5, 5, 10, 10, {colorSpaceConversion: "p3",
    //         resizeWidth: 20, resizeHeight: 20, resizeQuality: "pixelated"}),
    // ]).then(([noCropHigh, noCropMedium, noCropLow, noCropPixelated, cropHigh,
    //           cropMedium, cropLow, cropPixelated]) => {
    ]).then(([noCropHigh]) => {
        checkNoCrop(noCropHigh, colorInfo, sourceType);
        // checkNoCrop(noCropMedium, colorInfo, sourceType);
        // checkNoCrop(noCropLow, colorInfo, sourceType);
        // checkNoCrop(noCropPixelated, colorInfo, sourceType);
        // checkCrop(cropHigh, colorInfo, sourceType);
        // checkCrop(cropMedium, colorInfo, sourceType);
        // checkCrop(cropLow, colorInfo, sourceType);
        // checkCrop(cropPixelated, colorInfo, sourceType);
        // // Brute-force comparison among all bitmaps is too expensive.
        // // In case of SVG, resize quality does not affect the images, so all
        // // of them are the same and the tests fail. Since, we ignore this test
        // // set for SVG.
        // if (sourceType != 'svg') {
        //     compareBitmaps(noCropHigh, noCropMedium);
        //     compareBitmaps(noCropLow, noCropPixelated);
        //     compareBitmaps(cropHigh, cropMedium);
        //     compareBitmaps(cropLow, cropPixelated);
        // }
    });
}

function testImageBitmapTransparent(source)
{
    return testImageBitmap(source, 'transparent', 'general');
}

function testImageBitmapFromVideo(source)
{
    return testImageBitmap(source, 'fullColor', 'video');
}

function testImageBitmapOpaque(source)
{
    return testImageBitmap(source, 'opaque', 'general');
}

function testImageBitmapFromSVG(source)
{
    return testImageBitmap(source, 'opaque', 'svg');
}

function initializeOffscreenCanvas(canvasColorSpace, canvasPixelFormat)
{
    var canvas = document.createElement("canvas");
    canvas.width = 2;
    canvas.height = 2;
    var offscreen = canvas.transferControlToOffscreen();
    var ctx = offscreen.getContext('2d');
    ctx.fillStyle = "rgba(155, 27, 27, 1)";
    ctx.fillRect(0, 0, 1, 1);
    ctx.fillStyle = "rgba(27, 155, 27, 1)";
    ctx.fillRect(1, 0, 1, 1);
    ctx.fillStyle = "rgba(27, 27, 155, 1)";
    ctx.fillRect(0, 1, 1, 1);
    ctx.fillStyle = "rgba(27, 27, 27, 1)";
    ctx.fillRect(1, 1, 1, 1);
    return offscreen;
}

function initializeOffscreenCanvasTransparent(canvasColorSpace, canvasPixelFormat)
{
    var canvas = document.createElement("canvas");
    canvas.width = 2;
    canvas.height = 2;
    var offscreen = canvas.transferControlToOffscreen();
    var ctx = offscreen.getContext('2d',
        {colorSpace: canvasColorSpace, pixelFormat:canvasPixelFormat});
    ctx.fillStyle = "rgba(155, 27, 27, 0.5)";
    ctx.fillRect(0, 0, 1, 1);
    ctx.fillStyle = "rgba(27, 155, 27, 0.5)";
    ctx.fillRect(1, 0, 1, 1);
    ctx.fillStyle = "rgba(27, 27, 155, 0.5)";
    ctx.fillRect(0, 1, 1, 1);
    ctx.fillStyle = "rgba(27, 27, 27, 0.5)";
    ctx.fillRect(1, 1, 1, 1);
    return offscreen;
}

//OffscreenCanvas - Opaque SRGB
promise_test(function() {
    var offscreen = initializeOffscreenCanvas('srgb', '8-8-8-8');
    return testImageBitmapOpaque(offscreen);
}, 'createImageBitmap in P3 from an opaque SRGB OffscreenCanvas with resize.');

//OffscreenCanvas - Opaque SRGB
// promise_test(function() {
//     var offscreen = initializeOffscreenCanvasTransparent('srgb', '8-8-8-8');
//     return testImageBitmapTransparent(offscreen);
// }, 'createImageBitmap in P3 from a transparent SRGB OffscreenCanvas with resize.');

////////////////////////////////////////////////////////////////////////////////

</script>
